backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose
bcrypt
passlib
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-ollama
langgraph
faiss-cpu
pymupdf
sentence-transformers
tavily-python
httpx
python-multipart
